"2:35pm. Let me do some writing. While I am in the grip of this obsession, I might as well go as far as I can with it.
There is something I do not understand.

How would a situation like...

```
forall a b. a -> b -> forall c. Constraint a b c -> c
```

...be handled.

It seems clear, but suppose you do apply the first few arguments...

```
forall c. Constraint a b c -> c
```

I am confused as to what exactly should the type of this be.

But well, is this any different from regular application?

2:40pm. Well, enough of this. How about I actually do some Haskell? I might as well check out how it handles its inference.

2:55pm. Now I am experiencing the joy of Haskell tooling once more. It seems like nothing change with Haskell in that regard for the last 3 years. It is still shit.

Ah, forget it. I do not feel like it right now.

What I want to do is think more about this.

3:10pm. This is fun in a way. I am not sure whether this is a good way to spend time, but by the end of this you can bet I am going to plan out how to do ad-hoc polymorphism regardless of whether I decide to actually do it or not.

Right now I am brainstorming how to type infer recursive functions.

There is just too much cruft.

3:25pm. I fiddling trying to get the Haskell to work in either VS Code or Atom. So far no luck. Piece of crap.

3:30pm. Agh, it would be simple to do regular HM inference to an extent.

Right now I am wondering - what would I do with an implicit forall that is never used.

Value restriction is the answer.

3:50pm. No, trying to figure out how to do unification on a language that HKTs, higher rankeds and typeclass constraints is just too much for me.

Since I haven't even started, it might be better for me to reason out a complete design with the assumption of just doing typeckecking. I think that if all the annotations were there, I could do a proper job of it.

5:40pm. I've been thinking. Yes.

This is working for me now. Though I want to do as much inference as possible automatically, putting in all these advanced type system features while maintaining F#/Haskell tier of inference is just too much for me.

It is better to start with full typechecking and then work my way towards implementing bidirectional inference and unification after that.

5:45pm. I really know nothing about these subjects, but my past 6+ months of experience in theorem proving has really familiarized me with type theory.

I have a decent feel for it now.

5:55pm. In order to make it useful, I really have to do plenty of scaffolding for the language first, and that should be my focus.

Getting all the difficulties out of the way is something that gives me hope.

I have a workable plan now.

I am going to have to study typeclasses more in depth, as well as how to do bidirectional type checking, but that is that. I'll slowly put the pieces together and by the end of this I will be an expert.

6pm. I do not think that dependent types will be necessary for the kinds of things I want to do - that work should be left for the agents. For myself, Spiral with top down type inference will suffice.

I am just going to make the language that I so painstaikingly crafted in 2017 and 2018 fully bloom.

I want to pursue something that will give me a feeling of gain, and that is certainly not math.

If I play this right, I do not think this should take me nearly as long as the v0.09 did. Six months should be more than enough to get everything under wraps, and 2-3 is a more realistic estimate.

By this point I am pretty familiar with various aspects of PL development. I am not the novice I once way.

Let me stop here.

There last few days have been incredible. I do not usually spend entire days in a row fully immersed in creative thinking of this sort. Not at this level of intensity."