"11:10am. I got up like 10m ago. Lately my sleep has been taking nearly 2/3rds of my day. And when I am both up and down, I think.
I've made a significant discovery. While thinking about how to resolve the ordering issues with typeclasses not only have I figured that out, but I realized something huge as a result. I now know how to eliminate all the performance issues with partial evaluation with regards to join points. In v0.1 I added objects, but I now know a much better solution that will render them obsolete.

11:20am. That great idea is to stratify the language into the top-level and the rest.

Instead of having the language run from top to bottom, I will instead let it have the `main` which will be the last function in the last module. That means that nothing will be run until the end and so there is no danger of adding typeclass instances anywhere. By the time the program really starts running, they will all be in the global dictionary anyway.

Because of typeclass constraints, I can't just run things arbitrarily anymore anyway - the main function itself will need to have constraint solving be done in order to run it in the first place. So it makes sense that it should be at the end.

11:30am. Perfect. Everything is falling perfectly into place.

11:50am. Today I feel like thinking as well. The despair is my friend here. It will drive me forward.

Before I start work on anything it is fine if I just imagine for a while what the new language will be like.

12:10pm. Yeah, it all fits now. Which is great.

But I am going to have to think whether I really want to do this at all.

I said 2-3 months, but that is just for the basic prototype. To go in this direction means to wholeheartedly dedicate myself to PL design. If I do this now, it is likely I will work on Spiral for years afterwards.

12:20pm. Do I really want to do this? Would I much rather take years in writing my own language, rather than install Linux so I can try out Deep CFR?

Do I really want to do in on my own that much? There are all sort of compromises that I can take here.

12:30pm. I might as well eat while I think. Let me take a break here for a while."