"10:10am. I am thinking through some design considerations. I've already done quite a bit.
10:20am. Ah, yes, let me report that internal Agda error from yesterday.

10:55am. I need just a little bit more and then I will have completely worked out the design for the new Spiral.

11am. Right now I am thinking how to design the raw Spiral. I've mostly settled on what I want in the fully typed version, but I need to think how those design considerations would affect its very close relative.

12:35pm. This kind of intense thinking mindset is good if you want to go on a diet. You just stop caring about things like hunger.

I am pretty much there with Spiral. This is going to be a radical redesign. v0.2 is going to be an entirely new language in more ways than one.

1:20pm. Done with breakfast. I'll be going out in a bit to give my dad a hand at work, but not that this will make any difference as I am in full thinking mode.

I really feel drained mentally right now despite not moving a finger.

I can really feel the new language taking shape in my mind.

I guess I really will do it in the end. I really thought I was done with language work, but I guess it is turning out to be my calling.

1:35pm. Can I believe this will be great? Can I really try to exceed languages like Haskell and F# in elegance.

Do I want to be a great programmer. A great programmer would do something like this without hesitation. He would think that this is not a huge undertaking, but the most basic of basics.

How about I forget what came before, bury my growing fears and just move forward?

That is the way to live.

Is there anything better programming wise than doing something like this? Is there anything else, I would rather be doing?

Spiral is just not enough. Type inference is something that needs to be automated. If I want to grow, I need to keep working at it.

Damn those CFR papers. They really kicked my ass.

I need to do what I did in 2017 and 2018 all over again, except this time better.

I am going to keep thinking about the design more, take a break for a few days if that is even possible for me at this point and then just go for it."