"9:05am. Merry Christmas. Tonight rather than obsessing over how to implement various language features I am more interested in determining what I can leave out. And what I will use the language for.
I was too cocky. As I said, that discussion I wrote yesterday was not how I really feel.

Partial evaluation in the context of functional programming is really important, but keeping programming straightforward is important as well. The old Spiral was fun and exciting, and I want to be the new Spiral to be as boring as possible.

It really bothers me how in as I wrote the outline yesterday I got into the mood of competing with Haskell. I do not need typeclasses. I do not need union types. I do not need to think what sort of features I can add to make type checking more just-in-time.

I wanted to get on Patreon and get money! I wanted to impress people!

9:15am. No!

All that I need is a clear goal and then start programming. Once you have your own course, you stop caring about such trivial things such as money and other people.

What is important is that I write the kind of code that I want to write.

9:20am. https://www.youtube.com/watch?v=utyBNDj7s2w
Bidirectional Type Checking

I realized halfway that this talk is really good, but I was too tired to pay attention yesterday. It would not be a bad idea to get a bit more familiar with things other than unification. Right now, what is bugging me is that I saw in some paper another method, but I have no idea what it is.

Let me finish this song and then I will watch the talk again. Then I will either do some research on type inference or start figuring out to open that new Spiral repo."