"10:25am. I am up. I've been in thinking mode again.
10:40am. ">>11252180
I want to say math, but given that literally no ML paper would ever hesitate to use 'reals', I am going to say computer science. Trying to understand neural networks using 'real' math is like trying to do number theory without knowing how integers are constructed. It is not going to work. An indication that this is true is the absolute dominance of empirical methods at this moment. The theorems never say anything about the real life performance of the algorithms and are nothing better than ad-hoc rationalizations.

A lot of the 'math' that goes into the field will need to be redone in order to make progress.

This would be so much work that I am not sure if just trying things out at random and seeing what works will get us to where want faster.

The entire field of ML is a joke right now. I thought it was the fault of the field itself, but the math community as a whole is to blame for not coming up with proper foundations to support it."

It just goes to show how pissed I am at the fakes that I am actually posting these screeds on /sci/.

Yesterday I had a lot of inspiration and I made the claim that I know how to do typecase, but mentally going through various examples revealed kinks in that conception.

10:45am. I was wrong. I definitely need to think more about this.

11am. Ah, crap. The more I think about this the worse it gets.

It made sense in the simple examples I had in mind yesterday, but now that I am considering how to fold over tuples and records I am quite confused. Yesterday it was simple, but now I am realizing that to typecase properly, it is not enough to just do the thing, I need to do it over the constraint as well. But then how would I deal with the recursive constraint? I actually have that in mind, but then what about the argument to the fold. It might get matched on. What do I do about that.

The issue is that I am now dealing with constraints, type functions and the value level. This is too complex.

Also because raw Types are unbounded, what I am trying to do might be even harder than dependent typing. Ah, crap.

11:15am. Ok...

First of all, let me consider the constraints as just being regular types...

10:35am. Ogh, damn it.

https://www.youtube.com/results?search_query=haskell+type+family&page=&utm_source=opensearch
https://www.youtube.com/watch?v=tJNU1H9XewM
Making sense of the Haskell type system by Ryan Lemmer at FnConf17

Let me see if Haskell gives me any inspiration. I am dead ending here.

11:55am. I am just spinning my wheels here. Let me have breakfast. I hadn't even started that talk. Now I am in thinking mode again. Damn it."