"1:10pm. Let me watch the talk. I'll go over the issues with my ideas later.
3:15pm. I wish I could say I am slacking, but I am just staring at the screen while thinking of how I could type "those" things!

It is too difficult. Some of the things which are very easy in raw Spiral would be downright impossible in the typed one or otherwise require very advanced features.

I thought I had great insight yesterday, but the really big idea was that I can run and evaluate any type level function. I can in fact do typecase at the type level and then from that reflect it at the value level...except not really.

A good example of a function that is giving me massive trouble is the equality function. Let me just consider the case of pair equality.

```
inl rec (=) forall t. (a : t) (b : t) : bool =
    match a,b with
    | (a,as),(b,bs) -> a = b && as = bs
    | a,b -> a = b
```

Even this little fragment is troublesome. Suppose I write a constraint like...

```
type eq_cons t =
    typecase t with
    | a * as -> eq_cons a = bool & eq_cons as = bool
    | t -> if is_prim t then bool else error_type "Must be a primitive type"
```

Then I add it to the function.

```
inl rec (=) forall t. converges (eq_cons t) => (a : t) (b : t) : bool =
    match a,b with
    | (a,as),(b,bs) -> a = b && as = bs
    | a,b -> a = b
```

Now on the face of it, this makes sense.

`eq_cons t` is easy to evaluate on the function call - you just run it. And in the `=` function it acts as a premise that I am not passing things like functions, records or keyword arguments. Beautiful, simply beautiful. It is a great idea.

But!

```
    match a,b with
    | (a,as),(b,bs) -> a = b && as = bs
    | a,b -> a = b
```

How exactly do I prove that this converges to `bool`?

Now you might say this is easy as all the branches are `bool`, therefore it must be `bool`.

Ok, fine. But what about...

```
typecase t with
| t :> {} -> Record.foldback (fun k x next s -> s && next (a = b k)) a id true
| _ when is_keyword_arg t -> Keyword.foldback (fun k x next s -> s && next (a = b k)) a id true
```

Keyword arguments in particular are killing me here. It is no problem to pattern match on a particular instance of them...

```
match a with
| arg1: q arg2: w -> ...
```

...but it is an absolute struggle to figure out what their parametric type would be.

Parametricity is great when talking about the very generic and the very specific, but it is really poor at the things in the middle.

3:40pm. I am dying here. I am really dying thinking how I might type this.

And the type might be more complex.

Something like...

```
type f_type t =
    typecase t with
    | a: _ | {qwe : _ } | int * int -> int
    | b: _ | {asd : _ } | float * float -> float

let f forall t. converges (f_type t) => (a : t) (b : t) : f_type t = ...
```

Suppose I have some match expression in the body of `f`. Then it really won't be so easy to prove that the thing holds.

What is easy? Typechecking this during partial evaluation. It really would not be difficult to make sure that the return type holds, as everything would be completely reduced. During partial evaluation there would be no unapplied type variables - only principal types.

Similarly during the function call to `f`, any term passed to `f_type` can be checked for convergence. If you pass something not allowed, you get a type error, and if you pass in a type variable, then you need to make sure that the caller has the constraint on it.

Similarly if you are doing type inference using constraint solving, then it is not a big deal to solve the constraint last after all the metavariables have been substituted for either type variables or principal types.

But actually, checking the body of the function with that awesome return type. That is fucking tough!

3:55pm. I think rather than break myself over it, I should just decide that at some point I need to drop down into uninferred Spiral and get a little dirty.

Yesterday I had a dream that I should be able to typecheck everything, but that was ephemeral. What I really can only do is type everything as types are their own untyped language.

4pm. I really need this. I can't just use `forall a b. a -> b` whenever I run into something I can't fully type. I can't be expect to provide `b` every time. That would put the burden of typing way on the runtime.

In fact, I should never use `forall a b. a -> b` as a type signature.

Those can only be exceptions.

4:10pm. I am dead here. Ok, so this wasn't as bad as the previous week, but I needed to work this out. Since the year is nearly over, I think I should start writing the two reviews. I meant to start work, before I got sidetracked, but that is what happens when you are still in the design stages. You can only start when you actually work out everything.

I might as well do it now, since it is not like I am going to be productive at this time. I can only hope I find some focus in January.

I really will. I will finish the Spiral project in 2020. I will bring the language to its full potential and get back to playing with ML and math again.

4:20pm. Let me do the chores here. It seems I will be Thinking about this till tomorrow at least."