"4:55pm. Done with chores. I do not really feel like starting the reviews right now. I'll do it tomorrow when I am fresh.
How disappointing. But if it were possible to get everything I would want in a language just like that, there would have already been a language that does all that I want.

If it was that easy, you can bet the Haskell guys would have done it.

5:10pm. Damn, I really can't think of a consistent set of rules for the partial function convergence. If I add that feature, do I really have to make that check on every type level function application? I think I do.

How complicated.

Explicit constraints are easy to propagate, but I am really unsure how to deal with this.

5:20pm. Gah, just a little bit more. If I can add at least this, then even if I cannot necessarily typeckeck, I will at least have a type system that can encompass all of Spiral.

...Well, all of the important stuff anyway.

5:30pm. Ah, it will come to me. Let me stop here for the day. Just agonizing about it the entire day will not solve anything.

The way I should do this is as I've said I would - do it using bidirectional typechecking at the beginning. Do all the type application and top level annotations explicitly. Put in the constraints explicitly. Then after that I will be able to figure out how to start inferring things automatically.

I mean this is hard, but it is also new. I am betting it would be a lot easier if I had some experience with it. And that is what I need to get.

Tomorrow, once again while I write the reviews I will try to gather determination. Yesterday's idea dispersed it, but I should be able to start for real with these insights.

I need to start focusing on this.

5:35pm. Much like the last few years, I expect 2020 to be entirely lackluster, but that does not mean I will be sitting still.

I will be perfecting what I started. And hopefully that will make room for the bigger and better things ahead."