"2:30pm. I need to slack a bit more. Sorry about that. Right now I am thinking about fundeps.
The issues of how to get more expressive types was something that had been torturing me for the last 10 days, and now a complete and proper solutions without any drawbacks has fallen into my lap. My mind is absolutely blown by this. I need some time to digest this properly.

2:35pm. I might as well do the chores now.

3:10pm. I am back. Now let me slack properly.

Right now I am thinking about typeclasses.

Typeclasses are pretty complicated and are not very Spiralic. I feel it would be a challenge to implement them. So either way they are going to come last, even after union types.

Let me add a bit more to the review.

"One particularly brilliant idea that I have which will make implementation much easier is to make the language of constraints the uninferred Spiral language. The way constraints in Spiral v0.2 will work is that the constraint solver will just run them through the partial evaluator after dealing with equalities. If the function finishes running then the constraint will be discharged, but if the function raises a type error it will be caught and shown to the user just-in-time. And this will be wonderful. It is actually possible to go further than that with this idea. It is possible to instruct the constrain solver to run the uninferred function and extract the return type - which means that functions of type `forall a b. (FunDep a ~> b) => a -> b` when applied could have their return type inferred based on the constraint. That would allow very complex types to be easily derived that would be impossible to express in Haskell...or dependently typed languages for that matter.

That having said, in terms of type system features, what I will be aiming for is just one level above F# and that is it. That means - standard HM global type inference + (higher kinded) type level functions + (impredicative) higher ranked functions + functional dependencies + row polymorphism for records + basic union types + typeclasses. No GADTs or anything more complicated that would be hard to infer automatically or otherwise require an uniform representation of values at runtime. This is really everything I or any user will need in the Easy part. Beyond that - and I can speak from my experience, adding more type system features will actually significantly decrease productivity. With full dependent typing I'd estimate that factor to be between 10-50x. Theorems really are best when they come for free."

3:40pm. Yeah, I am having a lot of inspiration today. I think this one idea is what might make Spiral the dominant functional programming language of the future.

Now that I have this idea it makes me curious how Zig and other languages with static introspection would deal with this.

I never really looked into them.

4:05pm. Let me commit here. I am not sure if I want to continue with the intention of doing anything else for the day. Writing all of that and doing the brainstorming drained a lot out of me. I guess I will do a bit more research and then resume slacking."