"4:30pm. https://github.com/sponsors/andrewrk
This guy is working on this full time. I can't really imagine this happening to me, but whatever. I'll keep his model in mind.

https://help.github.com/en/github/supporting-the-open-source-community-with-github-sponsors/setting-up-github-sponsors-for-your-user-account

It seems Github has some kind of sponsorship scheme.

4:45pm. The same for Patreon. I am looking at it and it does not seems like there are any country restrictions.

https://help.github.com/en/github/supporting-the-open-source-community-with-github-sponsors/setting-up-github-sponsors-for-your-user-account#submitting-your-bank-and-tax-information

Actually, reading a bit more closely the restrictions here are just for that StripeConnect thing. I am guessing wire transfer should be allowed anywhere.

4:50pm. Ok, rather than saying I would do it on Patreon let me keep my options open. I'll change my wording a bit.

4:55pm. Enough of this. I'll revisit this in a month after I've gotten some work done on the language. I'd rather defer taking other people's people's money in this manner for as long as possible.

Before I even do this, I'd like to have something to show for it at least.

Also that Zig guy said he only quit his job because he was busy with other people's requests. Spiral on the other hand is unlikely to garner me anything anyway as I have no users. So it is hardly a priority.

If I am going to get anything from this it will be ways down the road.

5:55pm. Ah, I just realized something. If I decide to do these dependent constraints, then what am I going to do about nominal placeholder types?

I can't have them.

And if I do not have those, how exactly am I supposed to implement the Tensor type?

Should I allow typecase in the nominal type?

6pm. No, just no. That would be out of the question.

By necessity, the `var` operation must be inapplicable on those.

6:05pm. No, actually I have no choice. I am going to have to allow typecase in the definition for the nominal type.

For the sake of practicality, I can't really avoid partial functions at the type level.

6:10pm. Hmmmm...

Actually, it should be fine to have `real nominal` types. I'll allow partial type level functions. One of the great benefits of nominal types is that they ward off the typechecker from their insides.

Actually, rather than generating the types directly, I think that just like functional dependencies I will just run the partial evaluator on some real function and strip the type of its result. That way I could reuse `tensor_init` to both make the type and the tensor itself. Brilliant.

6:15pm. Ironically, these ideas I am suddenly discovering now where all type inference tricks I've used in Spiral v0.09 to do inference by hand so to speak.

To think they'd be useful here in this form.

6:20pm. But of course they would be. The things I discovered in my programming journey might be tricks, but they are real magic.

There might be a basis for what I wrote in the Spiral documentation about it being another branch that was completely overlooked by the mainstream.

6:25pm. Both standard `nominal` and `real nominal` types have their place in integrating the type system with the uninferred segments.

6:30pm. The typechecker will never have to ever see inside of them.

But to actually unpack `real nominal` types in the typed segment, I am definitely going to need functional dependencies as a prerequisite.

Everything is falling into place. If all the types have something behind them, that means I really will be able to guarantee that all types have the same structure, this is something that I would have lost with placeholder empty types that I'd originally planned.

This sort of solution is definitely Spiralic.

6:40pm. Let me do guess on how much code all this will take. The syntax rewrite and top level functions should not take much, but now having an explicit value level and all those new statements will take a bit, but it should not be too bad.

v0.1 is 4.5k if I remember correctly.

For the prototype without the typechecker and only the raw part, I think the additions should not amount to much more than it is currently. Probably it will end up between 5-6k. But the typechecker and the constraint solver, those should be 1-2k.

Maybe it won't be too bad, but I can expect that all the additions will at worst double the size of the language implementation.

6:45pm. I should be able to do better than that. I do get the feeling that I should be able to push the complexity down.

Though eventually when it comes to adding editor support, all bets will be off.

Well, let me just do it and then I'll see how things unfold after that."