"8:40am. I am up. Today I will finally start work on the language. There is a lot to go through before I even do the partial evaluator properly, let alone the typechecker. So work will be slow, but after a month or so I should have done a fair bit. Since it has been a while since I worked on Spiral, it will take me a while to get into the swing of things. Obviously, I am going to have to start with the parser.
As this will be a very significant redesign, I am going to have to rewrite everything from the ground up.

But before that, let me chill for a while.

https://mangadex.org/title/21707/houkago-assault-girls

This is pretty interesting actually.

9am. Yeah, past 6pm after the whole day is over I can't really enjoy manga like I can now. Now it is really great.

9:05am. For nearly two weeks now it was quite strenuous resolving all the issues with the new language design, and now that this has been resolved, I can rest more easily now.

9:50am. Ok, let me finally start. First thing on the order of business is to figure out how to open a fresh F# project from VS Code. Then I'll make a new repo.

I'll be doing this in VS Code as VS's retarded buggy autocomplete is driving me crazy. Every 2 years VS seems to be getting worse. I do not remember this being a problem in VS 2015.

9:55am. Right. Let me start."